Break safir.logging up into internal modules
Move the code for safir.logging out of __init__.py into internal
modules and turn __init__.py into only imports and re-exports.

The logger_name global variable that is set by configure_logging
and consumed by the logger dependency no longer worked as expected
once code was refactored into submodules due to the well-known
deficiencies in Python imports of simple variables. This variable
was intended to be internal to Safir and should never have been
used by clients (no uses appear in the GitHub lsst-sqre org). Move
it to the logger dependency module, which is more realistically
where it belongs, and stop exporting it to users of the library.
