+++ changelog.d/20240802_155256_rra_DM_45575.md
@@ -1,0 +1,3 @@
+### Bug fixes
+
+- `safir.logging` previously exported a `logger_name` variable holding the logger name configured via `safir.logging.configure_logging`. This variable was never documented and was not intended for use outside of the library. It is no longer exported, has been renamed, and is now private to the library.
